+++ actionpack/lib/action_dispatch/system_testing/test_helpers/setup_and_teardown.rb
@@ -4,14 +4,6 @@ module ActionDispatch
   module SystemTesting
     module TestHelpers
       module SetupAndTeardown # :nodoc:
-        def host!(host)
-          ActiveSupport::Deprecation.warn \
-            "ActionDispatch::SystemTestCase#host! is deprecated with no replacement. " \
-            "Set Capybara.app_host directly or rely on Capybara's default host."
-
-          Capybara.app_host = host
-        end
-
         def before_teardown
           take_failed_screenshot
         ensure